+++ packages/rs-dpp/src/data_contract/document_type/class_methods/mod.rs
@@ -6,6 +6,7 @@ use crate::data_contract::errors::DataContractError;
 use crate::ProtocolError;
 
 mod create_document_types_from_document_schemas;
+mod parse_property_reference;
 mod should_use_creator_id;
 mod system_properties;
 mod try_from_schema;
@@ -23,6 +24,25 @@ pub(crate) fn consensus_or_protocol_data_contract_error(
     #[cfg(not(feature = "validation"))]
     {
         ProtocolError::DataContractError(data_contract_error)
+    }
+}
+
+#[inline]
+pub(crate) fn consensus_or_protocol_data_contract_error_from_wrapped_protocol_error(
+    data_contract_error: ProtocolError,
+) -> ProtocolError {
+    #[cfg(feature = "validation")]
+    {
+        match data_contract_error {
+            ProtocolError::DataContractError(err) => ProtocolError::ConsensusError(
+                ConsensusError::BasicError(BasicError::ContractError(err)).into(),
+            ),
+            err => err,
+        }
+    }
+    #[cfg(not(feature = "validation"))]
+    {
+        data_contract_error
     }
 }
