+++ packages/rs-platform-version/src/version/dpp_versions/dpp_contract_versions/mod.rs
@@ -1,7 +1,8 @@
-use versioned_feature_core::{FeatureVersion, FeatureVersionBounds};
+use versioned_feature_core::{FeatureVersion, FeatureVersionBounds, OptionalFeatureVersion};
 pub mod v1;
 pub mod v2;
 pub mod v3;
+pub mod v4;
 
 #[derive(Clone, Debug, Default)]
 pub struct DPPContractVersions {
@@ -34,6 +35,7 @@ pub struct DataContractMethodVersions {
 pub struct DocumentTypeClassMethodVersions {
     pub try_from_schema: FeatureVersion,
     pub create_document_types_from_document_schemas: FeatureVersion,
+    pub parse_property_reference: OptionalFeatureVersion,
 }
 
 #[derive(Clone, Debug, Default)]
